[update] review 작성 부분 컴포넌트 개별 분리 util폴더에 저장  #160
